altered authContext to authProvider
