+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom'
-import { AuthProvider } from './context/auth-context'
+import { AuthProvider } from './providers/AuthProvider'
 import MainNavigation from './components/Navigation/MainNavigation'
 import Routes from './Routes'
