+++ src/Routes.js
@@ -1,8 +1,8 @@
 import React, { useContext } from 'react'
 import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom'
 
-import AuthContext from './context/auth-context'
-import AuthPage from './pages/Auth'
+import { AuthContext } from './providers/AuthProvider'
+import AuthPage from './pages/AuthNew'
 import MainNavigation from './components/Navigation/MainNavigation'
 import AttendeesPage from './pages/Attendees'
 import DataListPage from './pages/DataList'
