+++ src/components/Drinks/CustomizedMenus.js
@@ -9,7 +9,7 @@ import MoreVertIcon from '@material-ui/icons/MoreVert'
 import styles from './customized-menu.module.css'
 import DeleteIcon from '@material-ui/icons/Delete'
 import SaveIcon from '@material-ui/icons/Save'
-import AuthContext from '../../context/auth-context'
+import { AuthContext } from '../../providers/AuthProvider'
 import Dialog from '../Dialog/Dialog'
 import { AttendeeContext } from '../../providers/AttendeeProvider'
 import { DrinkContext } from '../../providers/DrinkProvider'
