+++ src/pages/DataList.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react'
-import AuthContext from '../context/auth-context'
+import { AuthContext } from '../providers/AuthProvider'
 import styles from './dataList.module.css'
 
 // component to display data. This component is not completed yet.
